End Screen and Upgrade Screen are now fullscreen
Significant changes were made to the handling of end and upgrade screens. The code was refactored to spawn these screens as fullscreen overlays, improving user experience. Additionally, a function was added to despawn the end screen when exiting it. Some redundant code related to safe_ui queries was commented out as part of this optimization process. Lastly, minor adjustments were made in state transitions for better game flow control.