Introduced Discord integration
A new DiscordPlugin has been added to handle the game's interaction with Discord. This includes setting up a client, connecting it, and checking for activity changes. The plugin uses the discord_rich_presence crate and is implemented as a Bevy plugin. An ActivityState struct has also been introduced to keep track of player's current status, what they're doing, whether their activity is an instance context like a match, and the start time of their activity. The code related to this functionality was moved from main.rs into its own module (discord.rs).