Added gamepad support for Dash, Accept, and Glide actions
The input map has been updated to include gamepad button mappings. The Dash action now supports the West button on a gamepad, the Accept action can be triggered by the South button, and the Glide action is mapped to the North button. This enhancement provides an alternative control scheme for players using a gamepad.