Refactored dash skill into separate module
The dash functionality has been refactored and moved into a new, separate module. This includes the creation of a new `SkillPlugin` and `DashPlugin`. The dash cooldown function has also been relocated to this new module. The main file was updated accordingly to reflect these changes, with the removal of the old dash code and addition of the new plugin.